pyproject.toml
[project]
name = "langgraph-agent-app"
version = "0.1.0"
description = "Langgraph Agent App"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-openai",
    "GitPython",
    "pandas"
]
